seperated itemfilter stories
